#6 Activity Controller tests.

+++ src/Controller/ActivityController.php
@@ -25,6 +25,8 @@ class ActivityController extends AbstractController {
 
     /**
      * @Route("/admin/activities", name="activities")
+     * @param Request $request
+     * @return Response
      */
     public function indexAction(Request $request)
     {
@@ -36,8 +38,11 @@ class ActivityController extends AbstractController {
 		]);
     }
 
-	/**
+    /**
      * @Route("/admin/activity/new", name="activity_new")
+     * @param Request $request
+     * @param TranslatorInterface $trans
+     * @return RedirectResponse|Response
      */
     public function newActivity(Request $request, TranslatorInterface $trans) {
 
@@ -73,6 +78,10 @@ class ActivityController extends AbstractController {
 
     /**
      * @Route("/admin/activity/edit/{id}", name="activity_edit")
+     * @param Request $request
+     * @param TranslatorInterface $trans
+     * @param $id
+     * @return RedirectResponse|Response
      */
     public function editActivity(Request $request, TranslatorInterface $trans, $id) {
 
@@ -80,7 +89,7 @@ class ActivityController extends AbstractController {
 
 		//Check if object exists
 		if(empty($object))
-			return $this->redirectToRoute('admin');
+			return $this->redirectToRoute('activities');
 
 		$options = ['submit'=>$trans->trans('Edit')];
 		$form = $this->createForm(ActivityType::class, $object, $options);
@@ -110,8 +119,12 @@ class ActivityController extends AbstractController {
 
     /**
      * @Route("/admin/activity/delete/{id}", name="activity_delete")
-	 *
-	 * @IsGranted("ROLE_SUPER_ADMIN")
+     *
+     * @IsGranted("ROLE_SUPER_ADMIN")
+     * @param Request $request
+     * @param TranslatorInterface $trans
+     * @param $id
+     * @return RedirectResponse
      */
     public function deleteActivity(Request $request, TranslatorInterface $trans, $id) {
 
@@ -121,30 +134,33 @@ class ActivityController extends AbstractController {
 		if(empty($object))
 			return $this->redirectToRoute('activities');
 
-		/*if($object->hasEditions()) {
-			$this->addFlash('notice','<p class="alert alert-danger">'.$trans->trans('not_possible_activity_has_editions').'</p>');
-			return $this->redirectToRoute('activities');
-		}*/
+        if($object->hasContacts()) {
+            $this->addFlash('notice','<p class="alert alert-danger">'.$trans->trans('not_possible_activity_has_contacts').'</p>');
+            return $this->redirectToRoute('activities');
+        }
 
 		$entityManager = $this->getDoctrine()->getManager();
 		$entityManager->remove($object);
 		$entityManager->flush();
 
-		$this->addFlash('notice','<p class="alert alert-success">'.$trans->trans('Activity successfully removed.').'</p>');	
-		
-		$referer = $request->server->get('HTTP_REFERER');
-		return new RedirectResponse($referer);
+		$this->addFlash('notice','<p class="alert alert-success">'.$trans->trans('Activity successfully removed.').'</p>');
+
+        return $this->redirectToRoute('activities');
     }
 
     /**
      * @Route("/admin/activity/{id}", name="activity_id")
+     * @param Request $request
+     * @param TranslatorInterface $trans
+     * @param $id
+     * @return RedirectResponse|Response
      */
     public function singleActivity(Request $request, TranslatorInterface $trans, $id) {
-		$object=$this->getDoctrine()->getRepository(Action::class)->find($id);
+		$object=$this->getDoctrine()->getRepository(Activity::class)->find($id);
 
 		//Check if object exists
 		if(empty($object))
-			return $this->redirectToRoute('admin');
+			return $this->redirectToRoute('activities');
 
         return $this->render('default/activity.html.twig', array(
 			'title' => $trans->trans('Activity'),
